@@ -6,14 +6,17 @@
         "build:prod": "webpack --mode production --output-filename material-ui-icons.min.inc.js"
     },
     "dependencies": {
-        "@material-ui/icons": "4.4.1"
+        "@mui/icons-material": "5.8.4"
     },
     "devDependencies": {
-        "webpack": "4.39.3",
+        "webpack": "4.43.0",
         "webpack-command": "0.5.0",
-        "@babel/runtime": "^7.4.4",
-        "@material-ui/core": "^4.4.1",
+        "@babel/runtime": "^7.17.2",
+        "@mui/material": "^5.0.0",
         "react": "*",
-        "react-dom": "*"
+        "react-dom": "*",
+        "@types/react": "*",
+        "@emotion/react": "*",
+        "@emotion/styled": "*"
     }
 }
